geofence integration fix
This commit should not be useful, but proposing it anyway, because it
should address (IMHO) a programming error in GeoServer. This only meant
to harden things a little bit.

I suspect the code that is using reflection mechanisms to browse every
interfaces that the object respond to. But it actually only browse the
one of the final class. This is not a problem when dealing with
*InfoImpl objects, because they implements the expected interface
directly. But this is not the case for Secured*Info objects, because
they inherit from a base class (Decorating*Info) which actually
implements the expected interface.

Hence, different calls to clazz.getInterfaces() does not return the
expected interfaces. Replacing it by spring reflection helpers to get
every interfaces a proxied object should address.

Tests: runtime tested.