Fix GEOS-7032 for 2.3.2 (still used in 14.01)
This is an adapted version (and a simpler one) of the fix proposed by
the GeoServer team. It has been made in a hurry, and some places might
have been forgotten, but at least the WFS parts (and the main revealed
by the upstream patch) should have been fixed.

Tested against WFS getfeature operation, before and after this patch.
difficult to say if it introduces regressions though.